services/backend/requirements.txt
aerich==0.7.1
asyncpg==0.27.0
bcrypt==4.0.1
passlib==1.7.4
fastapi==0.88.0
python-jose==3.3.0
python-multipart==0.0.5
tortoise-orm==0.19.2
uvicorn==0.20.0
asyncio
ijson
shapely
webdriver_manager
pytest
pytest-asyncio
pyjwt
httpx==0.23.3
black==23.9.1
isort==5.12.0
autoflake==1.7.8
pylint